moved datasources to appcomponent for daggermock limitations
