+++ app/src/main/java/es/npatarino/android/gotchallenge/di/AppComponent.java
@@ -5,6 +5,10 @@ import javax.inject.Singleton;
 
 import dagger.Component;
 import es.npatarino.android.gotchallenge.data.source.remote.EndPoint;
+import es.npatarino.android.gotchallenge.data.source.remote.JsonMapper;
+import es.npatarino.android.gotchallenge.domain.datasource.local.CharacterLocalDataSource;
+import es.npatarino.android.gotchallenge.domain.datasource.remote.CharacterRemoteDataSource;
+import es.npatarino.android.gotchallenge.domain.repository.GotCharacterRepository;
 import okhttp3.OkHttpClient;
 import rx.Scheduler;
 
@@ -13,6 +17,15 @@ import rx.Scheduler;
  */
 @Singleton @Component(modules = AppModule.class)
 public interface AppComponent {
+
+    //repository for DaggerMock limitations
+    GotCharacterRepository gotCharacterRepository();
+
+    JsonMapper gotCharacterJsonMapper();
+
+    //datasource
+    CharacterRemoteDataSource characterRemoteDataSource();
+    CharacterLocalDataSource characterLocalDataSource();
 
 
     @Named("executorThread") Scheduler executorThread();
