+++ app/src/main/java/es/npatarino/android/gotchallenge/di/AppModule.java
@@ -1,11 +1,20 @@
 package es.npatarino.android.gotchallenge.di;
 
+import com.google.gson.Gson;
+
 import javax.inject.Named;
 import javax.inject.Singleton;
 
 import dagger.Module;
 import dagger.Provides;
+import es.npatarino.android.gotchallenge.data.GotCharacterRepositoryImp;
+import es.npatarino.android.gotchallenge.data.source.local.CharacterLocalDataSourceImp;
+import es.npatarino.android.gotchallenge.data.source.remote.CharacterRemoteDataSourceImp;
 import es.npatarino.android.gotchallenge.data.source.remote.EndPoint;
+import es.npatarino.android.gotchallenge.data.source.remote.JsonMapper;
+import es.npatarino.android.gotchallenge.domain.datasource.local.CharacterLocalDataSource;
+import es.npatarino.android.gotchallenge.domain.datasource.remote.CharacterRemoteDataSource;
+import es.npatarino.android.gotchallenge.domain.repository.GotCharacterRepository;
 import okhttp3.OkHttpClient;
 import rx.Scheduler;
 import rx.android.schedulers.AndroidSchedulers;
@@ -18,6 +27,31 @@ import rx.schedulers.Schedulers;
 public class AppModule {
 
     private static final String END_POINT = "https://raw.githubusercontent.com/tonilopezmr/Game-of-Thrones/master/app/src/test/resources/data.json";
+
+    @Provides
+    @Singleton
+    public JsonMapper provideGotCharacterJsonMapper(){
+        return new JsonMapper(new Gson());
+    }
+
+    @Provides
+    @Singleton
+    public CharacterLocalDataSource provideCharacterLocalDataSource() {
+        return new CharacterLocalDataSourceImp();
+    }
+
+    @Provides
+    @Singleton
+    public CharacterRemoteDataSource provideCharacterRemoteDataSource(OkHttpClient okHttpClient, EndPoint endPoint, JsonMapper jsonMapper) {
+        return new CharacterRemoteDataSourceImp(jsonMapper, endPoint, okHttpClient);
+    }
+
+
+    @Provides
+    @Singleton
+    public GotCharacterRepository provideGotCharacterRepository(CharacterRemoteDataSource remoteDataSource, CharacterLocalDataSource localDataSource) {
+        return new GotCharacterRepositoryImp(remoteDataSource, localDataSource);
+    }
 
     @Provides
     @Singleton
