+++ app/src/main/java/es/npatarino/android/gotchallenge/di/modules/CharactersModule.java
@@ -1,20 +1,11 @@
 package es.npatarino.android.gotchallenge.di.modules;
 
-import com.google.gson.Gson;
-
 import javax.inject.Named;
 
 import dagger.Module;
 import dagger.Provides;
-import es.npatarino.android.gotchallenge.data.GotCharacterRepositoryImp;
-import es.npatarino.android.gotchallenge.data.source.local.CharacterLocalDataSourceImp;
-import es.npatarino.android.gotchallenge.data.source.remote.CharacterRemoteDataSourceImp;
-import es.npatarino.android.gotchallenge.data.source.remote.EndPoint;
-import es.npatarino.android.gotchallenge.data.source.remote.JsonMapper;
 import es.npatarino.android.gotchallenge.di.Activity;
 import es.npatarino.android.gotchallenge.domain.GoTCharacter;
-import es.npatarino.android.gotchallenge.domain.datasource.local.CharacterLocalDataSource;
-import es.npatarino.android.gotchallenge.domain.datasource.remote.CharacterRemoteDataSource;
 import es.npatarino.android.gotchallenge.domain.interactor.GetCharactersByHouseUseCase;
 import es.npatarino.android.gotchallenge.domain.interactor.common.GetListUseCase;
 import es.npatarino.android.gotchallenge.domain.repository.GotCharacterRepository;
@@ -22,7 +13,6 @@ import es.npatarino.android.gotchallenge.presenter.CharacterListPresenter;
 import es.npatarino.android.gotchallenge.presenter.CharacterListPresenterImp;
 import es.npatarino.android.gotchallenge.presenter.GotCharacterListByHousePresenter;
 import es.npatarino.android.gotchallenge.presenter.GotCharacterListByHousePresenterImp;
-import okhttp3.OkHttpClient;
 import rx.Scheduler;
 
 /**
@@ -32,29 +22,6 @@ import rx.Scheduler;
 @Module
 public class CharactersModule {
 
-    @Provides
-    @Activity
-    public JsonMapper provideGotCharacterJsonMapper(){
-        return new JsonMapper(new Gson());
-    }
-
-    @Provides
-    @Activity
-    public CharacterLocalDataSource provideCharacterLocalDataSource() {
-        return new CharacterLocalDataSourceImp();
-    }
-
-    @Provides
-    @Activity
-    public CharacterRemoteDataSource provideCharacterRemoteDataSource(OkHttpClient okHttpClient, EndPoint endPoint, JsonMapper jsonMapper) {
-        return new CharacterRemoteDataSourceImp(jsonMapper, endPoint, okHttpClient);
-    }
-
-    @Provides
-    @Activity
-    public GotCharacterRepository provideGotCharacterRepository(CharacterRemoteDataSource remoteDataSource, CharacterLocalDataSource localDataSource) {
-        return new GotCharacterRepositoryImp(remoteDataSource, localDataSource);
-    }
 
     @Provides
     @Activity
